Stateless: Add in some stateless support
Client able to send data for every FETCH command to get next Block2 response.

Server able to not keep state while sending Block2 responses back.

+++ include/coap3/coap_net_internal.h
@@ -188,7 +188,7 @@ struct coap_context_t {
   uint8_t mcast_per_resource;      /**< Mcast controlled on a per resource
                                         basis */
 #endif /* COAP_SERVER_SUPPORT */
-  uint8_t block_mode;              /**< Zero or more COAP_BLOCK_ or'd options */
+  uint32_t block_mode;             /**< Zero or more COAP_BLOCK_ or'd options */
 };
 
 /**
